requirements.txt
fastapi
uvicorn[standard]
langchain
langchain_ollama
weaviate-client
pydantic
pydantic-settings
sentence-transformers
httpx